fix: separate to components
